+++ spec/dummy/db/migrate/20140602115050_create_api_notify_tasks.rb
@@ -9,8 +9,11 @@ class CreateApiNotifyTasks < ActiveRecord::Migration
       t.text :response
       t.integer :depending_id
       t.boolean :done, default: false
+      t.string :changes_hash, index: true, limit: 32
 
       t.timestamps
     end
+
+    add_index :api_notify_tasks, [:api_notifiable_id, :api_notifiable_type], name: :api_notify_tasks_unique_index_on_api_notifiable
   end
 end