+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Changelog
 
+### 0.8.24
+* Added sidekiq.logger.warn when sidekiq_retries_exhausted
+* Added changes_hash field to tasks. It allows find unique unsumited tasks and prevent them for double executing.
+* Added LOGGER to setup_task method to inform when task is being created
+! By this changes you have to create new field for api_notify_tasks
+    add_column :api_notify_tasks, :changes_hash, limit: 32
+    add_index :api_notify_tasks, :changes_hash
+
 ### 0.8.23
 * Allow current request to execute parallel parent request, when using force_parent_sync. Downside is that it makes two requests when creating new child element with parent.